Fix code after review
[#807]
