+++ App/Controllers/Api/LibrariesController.php
@@ -4,22 +4,18 @@ namespace App\Controllers\Api;
 
 use App\Models\CacheInterface;
 use App\Models\Resource\LibrariesResource;
-use App\Models\Resource\LibraryResource;
 
 class LibrariesController extends AbstractApiController
 {
     private const CACHE_NAME = 'libraries';
     protected $librariesResource;
-    protected $libraryResource;
 
     public function __construct(
-        CacheInterface $cacheModel,
-        LibraryResource $libraryResource,
+        CacheInterface    $cacheModel,
         LibrariesResource $librariesResource,
-        $param = null
+                          $param = null
     ) {
         parent::__construct($cacheModel, $param);
-        $this->libraryResource   = $libraryResource;
         $this->librariesResource = $librariesResource;
     }
 
@@ -71,7 +67,7 @@ class LibrariesController extends AbstractApiController
             return false;
         }
 
-        $libraryModel = $this->libraryResource->getLibrary($this->param);
+        $libraryModel = $this->librariesResource->getLibrary($this->param);
 
         $data = $this->getLibrary($libraryModel);
         $this->updateCache(self::CACHE_NAME, $data);
@@ -83,7 +79,7 @@ class LibrariesController extends AbstractApiController
     private function createElement()
     {
         $data = $this->endCodeJson();
-        $libraryModel = $this->libraryResource->createLibrary($data['name'], $data['address']);
+        $libraryModel = $this->librariesResource->createLibrary($data['name'], $data['address']);
         header('Status: 200');
 
         $data = $this->getLibrary($libraryModel);
@@ -94,7 +90,7 @@ class LibrariesController extends AbstractApiController
     private function editElement()
     {
         $data = $this->endCodeJson();
-        $libraryModel = $this->libraryResource->editLibrary($data['name'], $data['address'], $this->param);
+        $libraryModel = $this->librariesResource->editLibrary($data['name'], $data['address'], $this->param);
         header('Status: 200');
 
         $data = $this->getLibrary($libraryModel);
@@ -104,7 +100,7 @@ class LibrariesController extends AbstractApiController
 
     private function deleteElement()
     {
-        $this->libraryResource->deleteLibrary($this->param);
+        $this->librariesResource->deleteLibrary($this->param);
         header('Status: 200');
 
         $this->cacheModel->clearCache(self::CACHE_NAME, true, $this->param);
