+++ App/Controllers/BookController.php
@@ -3,7 +3,7 @@
 namespace App\Controllers;
 
 use App\Blocks\BookBlock;
-use App\Models\Resource\BookResource;
+use App\Models\Resource\BooksResource;
 use App\Models\Resource\Environment;
 use App\Models\SessionModel;
 use App\Models\TokenModel;
@@ -15,15 +15,15 @@ class BookController extends AbstractController
         TokenModel $tokenModel,
         Environment $environment,
         BookBlock $block,
-        BookResource $resource
+        BooksResource $resource
     ) {
         parent::__construct($session, $tokenModel, $environment, $resource, $block);
     }
 
     public function execute()
     {
-        $bookResource = $this->resource;
-        $bookModel = $bookResource->getBook($this->getParam('id'));
+        $BooksResource = $this->resource;
+        $bookModel = $BooksResource->getBook($this->getParam('id'));
 
         $this->renderPage('book', $this->block, $bookModel);
     }
