+++ App/Controllers/DeleteAuthorController.php
@@ -2,7 +2,7 @@
 
 namespace App\Controllers;
 
-use App\Models\Resource\AuthorResource;
+use App\Models\Resource\AuthorsResource;
 use App\Models\Resource\Environment;
 use App\Models\SessionModel;
 use App\Models\TokenModel;
@@ -13,7 +13,7 @@ class DeleteAuthorController extends AbstractController
         SessionModel $session,
         TokenModel $tokenModel,
         Environment $environment,
-        AuthorResource $resource
+        AuthorsResource $resource
     ) {
         parent::__construct($session, $tokenModel, $environment, $resource);
     }
