+++ App/Controllers/DeleteBookController.php
@@ -2,7 +2,7 @@
 
 namespace App\Controllers;
 
-use App\Models\Resource\BookResource;
+use App\Models\Resource\BooksResource;
 use App\Models\Resource\Environment;
 use App\Models\SessionModel;
 use App\Models\TokenModel;
@@ -13,7 +13,7 @@ class DeleteBookController extends AbstractController
         SessionModel $session,
         TokenModel $tokenModel,
         Environment $environment,
-        BookResource $resource
+        BooksResource $resource
     ) {
         parent::__construct($session, $tokenModel, $environment, $resource);
     }
