+++ App/Controllers/LibraryController.php
@@ -2,11 +2,9 @@
 
 namespace App\Controllers;
 
-use App\Blocks\CategoriesBlock;
 use App\Blocks\LibraryBlock;
-use App\Models\Resource\CategoriesResource;
 use App\Models\Resource\Environment;
-use App\Models\Resource\LibraryResource;
+use App\Models\Resource\LibrariesResource;
 use App\Models\SessionModel;
 use App\Models\TokenModel;
 
@@ -17,7 +15,7 @@ class LibraryController extends AbstractController
         TokenModel $tokenModel,
         Environment $environment,
         LibraryBlock $block,
-        LibraryResource $resource
+        LibrariesResource $resource
     ) {
         parent::__construct($session, $tokenModel, $environment, $resource, $block);
     }
