+++ App/Controllers/PostCreateBookController.php
@@ -2,10 +2,8 @@
 
 namespace App\Controllers;
 
-use App\Blocks\CategoriesBlock;
 use App\Exceptions\MvcException;
-use App\Models\Resource\BookResource;
-use App\Models\Resource\CategoriesResource;
+use App\Models\Resource\BooksResource;
 use App\Models\Resource\Environment;
 use App\Models\SessionModel;
 use App\Models\TokenModel;
@@ -16,7 +14,7 @@ class PostCreateBookController extends AbstractController
         SessionModel $session,
         TokenModel $tokenModel,
         Environment $environment,
-        BookResource $resource
+        BooksResource $resource
     ) {
         parent::__construct($session, $tokenModel, $environment, $resource);
     }
