+++ App/Controllers/PostCreateLibraryController.php
@@ -2,10 +2,8 @@
 
 namespace App\Controllers;
 
-use App\Blocks\CategoriesBlock;
-use App\Models\Resource\CategoriesResource;
 use App\Models\Resource\Environment;
-use App\Models\Resource\LibraryResource;
+use App\Models\Resource\LibrariesResource;
 use App\Models\SessionModel;
 use App\Models\TokenModel;
 
@@ -15,7 +13,7 @@ class PostCreateLibraryController extends AbstractController
         SessionModel $session,
         TokenModel $tokenModel,
         Environment $environment,
-        LibraryResource $resource
+        LibrariesResource $resource
     ) {
         parent::__construct($session, $tokenModel, $environment, $resource);
     }
