+++ App/Controllers/PostEditAuthorController.php
@@ -2,9 +2,7 @@
 
 namespace App\Controllers;
 
-use App\Blocks\CategoriesBlock;
-use App\Models\Resource\AuthorResource;
-use App\Models\Resource\CategoriesResource;
+use App\Models\Resource\AuthorsResource;
 use App\Models\Resource\Environment;
 use App\Models\SessionModel;
 use App\Models\TokenModel;
@@ -15,7 +13,7 @@ class PostEditAuthorController extends AbstractController
         SessionModel $session,
         TokenModel $tokenModel,
         Environment $environment,
-        AuthorResource $resource
+        AuthorsResource $resource
     ) {
         parent::__construct($session, $tokenModel, $environment, $resource);
     }
