+++ App/Models/Resource/ModifyUserResource.php
@@ -19,8 +19,7 @@ class ModifyUserResource extends AbstractResource
         $password = $this->hashPassword($password);
 
         $query = 'UPDATE users set password = ?, name = ?, surname = ?, email = ? WHERE id = ?';
-        $db = $this->di->get(Database::class);
-        $db = $db->getConnection();
+        $db = $this->database->getConnection();
 
         $stmt = $db->prepare($query);
         $stmt->execute([$password, $name, $surname, $email, $id]);
@@ -29,8 +28,7 @@ class ModifyUserResource extends AbstractResource
     public function getUser(int $id): AbstractModel
     {
         $query = 'SELECT name, surname, email FROM users WHERE id = ?';
-        $db = $this->di->get(Database::class);
-        $db = $db->getConnection();
+        $db = $this->database->getConnection();
 
         $stmt = $db->prepare($query);
         $stmt->execute([$id]);
