+++ App/Router/Router.php
@@ -16,6 +16,13 @@ class Router extends AbstractRouter
 
         $uri = $this->handleUri($path);
 
+        $isMethodPost = $_SERVER['REQUEST_METHOD'] === 'POST';
+        $isMethodGet  = $_SERVER['REQUEST_METHOD'] === 'GET';
+        $isPostRegistration = stripos($uri, 'postRegistration');
+        $isCreate = stripos($uri, 'create');
+        $isEdit   = stripos($uri, 'edit');
+        $isLogin  = stripos($uri, 'postLogin');
+
         if ($uri == '') {
             return $this->di->get(Controllers\HomePageController::class);
         }
@@ -24,7 +31,7 @@ class Router extends AbstractRouter
             return $this->di->get(Controllers\RegistrationController::class);
         }
 
-        if ($_SERVER['REQUEST_METHOD'] == 'POST' && stripos($uri, 'postRegistration') !== false) {
+        if ($isMethodPost && $isPostRegistration !== false) {
             return $this->di->get(Controllers\PostRegistrationController::class);
         }
 
@@ -34,36 +41,36 @@ class Router extends AbstractRouter
         $class = 'App\Controllers\\' . $class;
 
         if (class_exists($class) && isset($sessionId)) {
-            if ($_SERVER['REQUEST_METHOD'] == 'GET' && stripos($uri, 'create') !== false) {
+            if ($isMethodGet && $isCreate !== false) {
                 return $this->di->get($class);
-            } elseif ($_SERVER['REQUEST_METHOD'] == 'POST' && stripos($uri, 'create') !== false) {
-                $class = explode('\\', $class);
-                $postName = 'Post' . array_pop($class);
-                $class[] = $postName;
-                $class = implode('\\', $class);
-
-                return $this->di->get($class);
+            } elseif ($isMethodPost && $isCreate !== false) {
+                return $this->di->get($this->getClass($class));
             }
 
-            if ($_SERVER['REQUEST_METHOD'] == 'GET' && stripos($uri, 'edit') !== false) {
+            if ($isMethodGet && $isEdit !== false) {
                 return $this->di->get($class);
-            } elseif ($_SERVER['REQUEST_METHOD'] == 'POST' && stripos($uri, 'edit') !== false) {
-                $class = explode('\\', $class);
-                $postName = 'Post' . array_pop($class);
-                $class[] = $postName;
-                $class = implode('\\', $class);
+            } elseif ($isMethodPost && $isEdit !== false) {
 
-                return $this->di->get($class);
+                return $this->di->get($this->getClass($class));
             }
 
             return $this->di->get($class);
         } elseif (!isset($sessionId)) {
-            if ($_SERVER['REQUEST_METHOD'] == 'POST' && stripos($uri, 'postLogin') !== false) {
+            if ($isMethodPost && $isLogin !== false) {
                 return $this->di->get(Controllers\PostLoginController::class);
             }
             return $this->di->get(Controllers\LoginController::class);
         }
 
         return $this->di->get(Controllers\Error404Controller::class);
+    }
+
+
+    private function getClass(string $class): string
+    {
+        $class = explode('\\', $class);
+        $postName = 'Post' . array_pop($class);
+        $class[] = $postName;
+        return implode('\\', $class);
     }
 }
